backend/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Env
python-dotenv==1.0.0

# Slack
slack-sdk==3.26.0

# LangChain - using compatible 0.3.x versions
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-groq==0.2.0

# Google APIs
google-generativeai==0.3.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# Multipart support for file uploads (FastAPI)
python-multipart==0.0.6

# Pydantic v2 - upgraded to satisfy langchain>=2.7.4 requirement
pydantic==2.7.4
pydantic-settings==2.4.0

# Async HTTP
aiohttp==3.9.1
async-timeout==4.0.3
httpx==0.24.1

# OAuth & Auth libraries

# Database / ORM
SQLAlchemy==2.0.20
asyncpg==0.27.0
psycopg2-binary==2.9.7

# Typing
typing-extensions==4.9.0

# Additional dependencies for proper functionality
tenacity==8.2.3
tiktoken==0.5.1
numpy==1.24.3
